Time: 16 ms (73.54%), Space: 44.1 MB (47.39%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -24,7 +24,7 @@ public class Solution {
 				int l = nums.length - 1;
 				while (k < l) {
 
-					long sum = nums[i] + nums[j] + nums[k] + nums[l];
+					int sum = nums[i] + nums[j] + nums[k] + nums[l];
 
 					if (sum == target) {
 						result.add(Arrays.asList(nums[i], nums[j], nums[k], nums[l]));
